# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication and account-security API: registration with email verification, JWT login with TOTP MFA, Google OAuth 2.0, session and device management, and passkeys/WebAuthn. The response envelope is {code, message, data} on every endpoint.",
        "title": "FinnApiGo Auth API",
        "contact": {
            "name": "FinnApiGo",
            "url": "https://github.com/NguyenQuan121321/FinnApiGo/issues"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/auth/mfa/methods": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Aggregates active MFA methods, remaining recovery codes, and default method for the caller.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MFA"
                ],
                "summary": "List MFA methods",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.MFAMethodsEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/mfa/totp/disable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Disables TOTP authentication for the caller. Requires either an active sudo token (X-Sudo-Token) or current password + (TOTP code or recovery code).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MFA"
                ],
                "summary": "Disable TOTP",
                "parameters": [
                    {
                        "description": "Password and TOTP/recovery code (optional if X-Sudo-Token provided)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.TOTPDisableRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.NullDataEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/mfa/totp/enable": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Begins TOTP enrollment. Returns the shared secret and provisioning URI. If TOTP is already active, stages a new secret (requires X-Sudo-Token for re-enrollment).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MFA"
                ],
                "summary": "Begin TOTP enrollment",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.TOTPEnableEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/mfa/totp/recovery-codes": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Re-displays saved recovery codes after verifying a current TOTP code. Also mints a short-lived sudo token for the regenerate endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MFA"
                ],
                "summary": "View recovery codes",
                "parameters": [
                    {
                        "description": "Current TOTP code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.TOTPCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.RecoveryCodesViewEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/mfa/totp/recovery-codes/regenerate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "SudoAuth": []
                    }
                ],
                "description": "Invalidates existing recovery codes and generates a new set. Requires the X-Sudo-Token minted by the view-recovery-codes endpoint (GitHub-style sudo mode).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MFA"
                ],
                "summary": "Regenerate recovery codes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.RecoveryCodesEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/mfa/totp/validate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validates a TOTP code for re-authentication on an active session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MFA"
                ],
                "summary": "Validate a TOTP code",
                "parameters": [
                    {
                        "description": "Current TOTP code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.TOTPCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.NullDataEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/mfa/totp/verify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Confirms TOTP enrollment with a current code. Activates TOTP and returns single-use recovery codes (displayed once).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MFA"
                ],
                "summary": "Confirm TOTP enrollment",
                "parameters": [
                    {
                        "description": "Current TOTP code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.TOTPCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.RecoveryCodesEnvelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.TOTPCodeRequest": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 6
                }
            }
        },
        "internal_handlers.TOTPDisableRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "swagger.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "validation error"
                }
            }
        },
...
        "swagger.MFAMethodsData": {
            "type": "object",
            "properties": {
                "defaultMethod": {
                    "type": "string",
                    "example": "totp"
                },
                "passkeysCount": {
                    "type": "integer",
                    "example": 1
                },
                "recoveryCodesRemaining": {
                    "type": "integer",
                    "example": 8
                },
                "totpEnabled": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "swagger.MFAMethodsEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/swagger.MFAMethodsData"
                },
                "message": {
                    "type": "string",
                    "example": "mfa methods"
                }
            }
        },
        "swagger.NullDataEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "operation successful"
                }
            }
        },
...
        "swagger.RecoveryCodesData": {
            "type": "object",
            "properties": {
                "recoveryCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "a1b2c3d4e5f6",
                        "g7h8i9j0k1l2"
                    ]
                }
            }
        },
        "swagger.RecoveryCodesEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/swagger.RecoveryCodesData"
                },
                "message": {
                    "type": "string",
                    "example": "TOTP enabled"
                }
            }
        },
        "swagger.RecoveryCodesViewData": {
            "type": "object",
            "properties": {
                "recoveryCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "a1b2c3d4e5f6",
                        "g7h8i9j0k1l2"
                    ]
                },
                "sudoExpiresInSec": {
                    "type": "integer",
                    "example": 900
                },
                "sudoToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                }
            }
        },
        "swagger.RecoveryCodesViewEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/swagger.RecoveryCodesViewData"
                },
                "message": {
                    "type": "string",
                    "example": "Recovery codes"
                }
            }
        },
...
        "swagger.TOTPEnableData": {
            "type": "object",
            "properties": {
                "provisioningURI": {
                    "type": "string",
                    "example": "otpauth://totp/FinnApiGo:john@example.com?secret=JBSWY3DPEHPK3PXP\u0026issuer=FinnApiGo"
                },
                "secret": {
                    "type": "string",
                    "example": "JBSWY3DPEHPK3PXP"
                }
            }
        },
        "swagger.TOTPEnableEnvelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/swagger.TOTPEnableData"
                },
                "message": {
                    "type": "string",
                    "example": "TOTP enrollment pending verification"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token. Send as \"Bearer \u003ctoken\u003e\". For /auth/mfa/login-verify this must be the short-lived mfa_pending token issued by /auth/login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SudoAuth": {
            "description": "Short-lived sudo token minted by POST /auth/mfa/totp/recovery-codes; required for privileged operations.",
            "type": "apiKey",
            "name": "X-Sudo-Token",
            "in": "header"
        }
    }
}